+++ R/gk.R
@@ -111,6 +111,80 @@ gk_w <- function(x,pvar,qvar,pervar,prodID, sample) {
 
 }
 
+#' Compute GK index for a window using the iterative method
+#' @keywords internal
+#' @noRd
+gk_w_iterative <- function(x, pvar, qvar, pervar, prodID, sample, tolerance = 1/1e12,
+                           maxIter = 100){
+
+  # list of time periods
+  pers <- sort(unique(x[[pervar]]))
+  # list of products
+  prods <- sort(unique(x[[prodID]]))
+  # total time periods
+  obs <- max(x[[pervar]]) - min(x[[pervar]]) + 1
+  # number of products
+  n <- length(prods)
+
+  # if matching requested, keep only products that occur through the whole window
+  if(sample == "matched"){
+    x <- windowMatch(x, pervar, prodID)
+  }
+  else {
+    # fill out the gaps from missing/new products with zeros.
+    # this makes some operations easier, but won't contribute to the calculations
+    x <- fillMissing(x, pvar, qvar, pervar, prodID, priceReplace = 0, quantityReplace = 0)
+  }
+
+  qtn <- matrix(x[[qvar]], nrow = obs, byrow = TRUE)
+  # sum quantity for each product for all time periods
+  qt <- colSums(qtn)
+  # quantity shares
+  stn <- matrix(NA, nrow = obs, ncol = n)
+  for(i in 1:n){
+    stn[,i] <- qtn[,i]/qt[i]
+  }
+
+  calcPi <- function(P){
+    pi <- numeric(n)
+    for(i in 1:n){
+      pi[i] <- sum(stn[,i]*(x[x[[prodID]] == prods[i],][[pvar]]/P))
+    }
+    return(pi)
+  }
+
+  calcP <- function(Pi){
+    pindex <- numeric(obs)
+    for(i in 1:obs){
+      temp <- x[x[[pervar]] == pers[i],]
+      pindex[i] <- (temp[[pvar]]%*%temp[[qvar]])/(t(Pi)%*%temp[[qvar]])
+    }
+    return(pindex)
+  }
+
+  # initial calculation of P
+  pi <- rep(1, n)
+  p <- calcP(pi)
+
+  # iterate to solve
+  diff <- rep(1, obs)
+  iter <- 0
+  while(any(abs(diff) > tolerance) && iter <= maxIter){
+    prevP <- p
+    pi <- calcPi(p)
+    p <- calcP(pi)
+    diff <- p - prevP
+    iter <- iter + 1
+  }
+
+  if(iter == 100){
+    warning("Convergence not reached for given tolerance. GK iterations stopped at maxIter iterations. Try increasing the tolerance or maxIter to reach convergence.")
+  }
+
+  return(as.matrix(p/p[1]))
+
+}
+
 #' Compute the Geary-Khamis index
 #'
 #' @param x A dataframe containing price, quantity, a time period identifier
@@ -127,6 +201,18 @@ gk_w <- function(x,pvar,qvar,pervar,prodID, sample) {
 #' @param splice the splicing method to use to extend the index. Valid methods are
 #' window, movement, half, mean, fbew, fbmw, wisp, hasp or mean_pub. The default is mean.
 #' See details for important considerations when using fbew and fbmw.
+#' @param imputePrices the type of price imputation to use for missing prices.
+#' Currently only "carry" is supported to used carry-forward/carry-backward prices.
+#' Default is NULL to not impute missing prices.
+#' @param solveMethod the method to use to solve for the quality adjustment factors
+#' and the price levels. "inverse" uses a matrix inverse operation, is much more efficient, but
+#' may not work if there are many missing observations.
+#' "iterative" iterates between the equations for the quality adjustment factors and price levels
+#' and is much slower, but can be used even when there are a large number of missing observations.
+#' @param tolerance the tolerance for the iterative solving method. Smaller numbers will produce more
+#' accurate results, but take more iterations. Default is 1/1e12, which may be a little larger
+#' than machine precision, given by .Machine$double.eps.
+#' @param maxIter the maximum number of iterations for the iterative solving method.
 #' @details The splicing methods are used to update the price index when new data become
 #' available without changing prior index values. The window, movement, half and mean splices
 #' use the most recent index value as the base period, which is multiplied by a price movement
@@ -137,6 +223,11 @@ gk_w <- function(x,pvar,qvar,pervar,prodID, sample) {
 #' right window length chosen. For example, if you have monthly data and want December
 #' of each year to be the base period, then the first period in the data must be December
 #' and the window must be set to 13.
+#'
+#' It is recommended to use the matrix inverse method of solving the GK equations (the default)
+#' because the performance difference can be significant. If the matrix inverse method does
+#' not work then switch to the iterative method. The tolerance and maximum number of iterations
+#' in the iterative method can be adjusted to balance performance and precision.
 #' @examples
 #' # compute a Geary-Khamis index with mean splicing
 #' GKIndex(CES_sigma_2, pvar = "prices", qvar = "quantities", pervar = "time",
@@ -151,7 +242,9 @@ gk_w <- function(x,pvar,qvar,pervar,prodID, sample) {
 #' Khamis, S. H. 1970. “Properties and Conditions for the Existence of a New Type of Index Number.”
 #' Sankhya: The Indian Journal of Statistics, Series B (1960-2002) 32: 81–98.
 #' @export
-GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice = "mean"){
+GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice = "mean",
+                    imputePrices = NULL, solveMethod = "inverse", tolerance = 1/1e12,
+                    maxIter = 100){
 
   # check that only valid splice methods are chosen
   if(!(tolower(splice) %in% c("mean", "window", "movement", "half", "fbew", "fbmw", "wisp", "hasp", "mean_pub"))){
@@ -162,6 +255,11 @@ GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice =
   colNameCheck <- checkNames(x, c(pvar, qvar, pervar, prodID))
   if(colNameCheck$result == FALSE){
     stop(colNameCheck$message)
+  }
+
+  # check valid solveMethods are given
+  if(!tolower(solveMethod) %in% c("inverse", "iterative")){
+    stop("Not a valid solveMethod.")
   }
 
   # check that the time period variable is continuous
@@ -180,6 +278,13 @@ GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice =
     stop("The window length exceeds the number of periods in the data")
   }
 
+  # apply price imputation
+  if(!is.null(imputePrices)){
+    switch(imputePrices,
+           "carry" = {x <- imputeCarryPrices(x, pvar, qvar, pervar, prodID)},
+           stop("Invalid imputePrices argument"))
+  }
+
   # sort the dataset by time period and product ID
   x <- x[order(x[[pervar]], x[[prodID]]),]
 
@@ -195,7 +300,10 @@ GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice =
   xWindow <- x[x[[pervar]] >= 1 & x[[pervar]] <= window,]
 
   # call gk_w on first window
-  pGK[1:window,1] <- gk_w(xWindow, pvar, qvar, pervar, prodID, sample)
+  pGK[1:window,1] <- switch(solveMethod,
+                            "inverse" = gk_w(xWindow, pvar, qvar, pervar, prodID, sample),
+                            "iterative" = gk_w_iterative(xWindow, pvar, qvar, pervar,
+                                                         prodID, sample, tolerance, maxIter))
 
   # use a splicing method to compute the rest of the index
   if(n > window){
@@ -224,7 +332,10 @@ GKIndex <- function(x, pvar, qvar, pervar, prodID, sample = "", window, splice =
       }
 
       # call gk_w on this window
-      new <- gk_w(xWindow, pvar, qvar, pervar, prodID, sample)
+      new <- switch(solveMethod,
+                    "inverse" = gk_w(xWindow, pvar, qvar, pervar, prodID, sample),
+                    "iterative" = gk_w_iterative(xWindow, pvar, qvar, pervar,
+                                                 prodID, sample, tolerance, maxIter))
 
       # splice the new datapoint on
       switch(splice,